Organize cartography rendering and tighten sample contracts.
Move overlay modules under lib/cartography, split the stellar overlay pane into vector and wormhole marker components, use a repo-root ion voltage contract fixture, and type sample-at entries as a discriminated union on both API and frontend.
