+++ packages/frontend/src/components/map-graph/useOverlayPaneSize.ts
@@ -1,0 +1,23 @@
+import { useEffect, useState } from 'react'
+
+/** Track React Flow pane size for absolute SVG/marker overlays. */
+export function useOverlayPaneSize(domNode: HTMLElement | null): { width: number; height: number } {
+  const [size, setSize] = useState({ width: 0, height: 0 })
+
+  useEffect(() => {
+    if (!domNode) return
+    let raf = 0
+    const ro = new ResizeObserver((entries) => {
+      const { width, height } = entries[0]?.contentRect ?? { width: 0, height: 0 }
+      cancelAnimationFrame(raf)
+      raf = requestAnimationFrame(() => setSize({ width, height }))
+    })
+    ro.observe(domNode)
+    return () => {
+      cancelAnimationFrame(raf)
+      ro.disconnect()
+    }
+  }, [domNode])
+
+  return size
+}
