+++ packages/frontend/src/lib/cartography/ionStormCloudOverlay.ts
