+++ packages/frontend/src/lib/cartography/rasterizeMapField.ts
