+++ packages/frontend/src/lib/cartography/scalarFieldGrid.ts
