+++ packages/frontend/src/lib/cartography/stellarCartographyOverlay.test.ts
