+++ packages/frontend/src/lib/cartography/stellarCartographyOverlay.ts
