+++ packages/frontend/src/lib/cartography/stellarCartographyTheme.ts
